requirements.txt
fastapi
uvicorn
pandas
scikit-learn
joblib
streamlit
requests
azure-storage-blob
